Put background job processing into endIteration(...).
The background job processing will thus not tunnel through the reduction anymore and
potentially block the master.
However, the symmetric exchanger hopefully performs better as
background jobs are processed before it is ensured that the receives
are complete.

The whole thing is likely more a question of where the neighbours of a heavy
rank will wait: In this iteration or the one after.

Not tunneling reductions should not make a difference for reasonably balanced grids.