Adding Michael Dumbsers Runge-Kutta Finite-Differencing "Toy code" implementation of CCZ4 to the UserCodes.
This code was written by Michael to demonstrate that we should be able to evolve FO-CCZ4 with classical Method of Lines.

Note that this is another instance of adding the FO-CCZ4 PDE to the main repository. We want to keep this PDE code closed source, however, for the runtime of the project. Similar examples are given in the benchmarking folders where somebody copied the CCZ4 PDE all over the place.