LoadBalancing: Update
Tried all tricks to not distribute cells on level 2 directly but the balancing
doesn't take the number of children of the first 3^d children into
account before determining which child to fork.

We thus have a 1/3^d chance to fork off the correct child if one child
of a cell is refined further.

We have to improve how we identify the critical
path in our tree.