Plotters: Implementing a general-purpose high level CSV writer.
It basically allows to map C structures onto CSV lines. It is kind of generic for extensions towards self-describing data. Once they are in place, the ReductionWriter classes should be ported to this CSVWriter. I did not do it now because they have still a feature to do global reductions with MPI_Send/MPI_Recv. Thought this turned out to be very very bad for ExaHyPE which struggles with MPI blockades so much, it's still there and in principle working.

Needless to say, there are various other CSV writers in place: There is even a directory plotters/CSV which collects ASCII/CSV-centric volumetric outputting (which would have been good for debugging three years ago but is nowadays not really of interest any more), and there are routines in Peanos tarch, but I never used them.

In our astrophysics applications, we need such kind of general-purpose CSV writers for any kind of time series data which could also be of various forms.