ADER-DG BoundaryConditions call: Passing the solution (luh) to the Abstract*Solver.
This allows BoundaryConditions to access all the patch values. This is handy for solving a different PDE on the boundary which is has the patch values as initial data.

This is an API change which naturally requires you to rebuild your toolkit, do an "rm Abstract* *.o" in your application and rerun the toolkit in your application to get it back running.

In case you changed your Abstract*Solver by hand, you can easily insert the neccessary changes by inspecting the diff of the ADERDGSolver.{cpp,h} of this commit, or you move your changes to a safe place, regenerate the files and diff them against your changes to pick up the new lines.

Having said this all; this is not a major API change but only a small addition.