Kernels: Adding a small standalone RKDG integrator.
The code is quite general and fits nicely to the ExaHyPE solver concept, therefore I put it into the kernels namespace.

However, it is not a full PDE solver as it lacks any boundary exchange. It is suitable only for integrating a single DG cell in time.

I actually use it for integrating a PDE at the boundary.