Toolkit2: A[n unfinished] parser for the original specfication file format.
This is work in progress. Tokenization and comment stripping works, but the mapping on the JSON-Schema is still theoretical code (=untested).

The idea is however to embed the lookup rules for the values right into the JSON-Schema of the new format. This allows to introduce new variables easily at a single place (=the schema) without the need to update the parser for the old format. It is also kind of elegant and readable and allows to associate old and new format more easily.

However, in its current form it is not suitable "the other way around", i.e. to generate an old specfile from JSON. To do so, I have a Jinja template for a specfile, however I won't put it here into the Toolkit2 because we should not need it for the way how the Toolkit is supposed to work.

This commit is only for collaborative purposes, the code is not yet working. Hopefully the JSON-Schema still is, thought.