Toolkit2: Adding an "omni-reader" for YAML, Human JSON, MEXA, ExaHyPE-v1.
We could even add a reader for Dominic's favourite "INI" format.

From my POV, the idea is that the toolkit has some CLI similar to git (checkout mexa.py --help to see what I mean), i.e. with subcommands which allow to use these different file formats. Or we provide several entrance points instead of a single one. It is a common understanding so far that users should not really be forced to write JSON, it is just inappropriate as it is very strict and does not allow comments.