Opening Pandora's Box: Outsourcing the ExaHyPE parser to Python.
This is a big step discussed between a few of the committee people in order to support ExaHyPE-v1 "original format" specification files as "first class citizens" without having the need to convert them in a preprocessing step.

That is, in this edit I basically replaced the need to call the new ExaHyPE not as

  ./ExaHyPE-MyPDE <(toolkit.sh --validate-only Classic.exahype)

but as

  ./ExaHyPE-MyPDE Classic.exahype

and the ExaHyPE code calls the Toolkit for conversion to JSON internally (degrading JSON as an inter-process serialization/communication protocol) at startup. A nice by-product is that we can make a JSON-Schema validation without any hazzle. And that we can support virtually any file formats -- there are readers for all kind of exotic nested-dictionary-readers already in the exahype.specfile python namespace in the new toolkit.

This commit is a "proof of concept", it works with the EulerFlow specfiles.

However, I have not tested running ExaHyPE in depth. Especially right now, I see that I correctly pass JSON to the ExaHyPE Parser (which no more deserves the name "Parser", actually), but subsequently there is a lot of parameter hazzle and stuff seems to miss.