Toolkit2: Worsening. Please review.
I tried to run the JSON example specfiles but they did not work for me. So I edited the Schema (probably worsened it). And then removed stuff from the JSON specfile. Sorry for that.

Subsequently, I made edits to the toolkit meta shellscript. The "exec" line is important, but I removed dependencies because they were not fulfilled on my system and I don't see why we need them for the toolkit (maybe for the optimized kernels, but JM prefers to keep them seperated, apparently, anyway).