Toolkit2 specfile1 reader: Fixing passing selection strings.
I acknowledge the select statement should actually be part of the parameters, but somehow the naming is damned anyway. The proper way of doing this is to put "select" on an equal footing as any other parameter, *not* nesting arbitrary parameters into the "parameters" section but keeping them all as parts next to "variables", "start", "every", etc.

Anyway, this solves #249.