Reduced number of heap accesses during Riemann solves.
Further loop fusion might be necessary to
reduce the time spent in the Riemann solves.

The Riemann solves slow down the production
of jobs for consumer tasks if highly
adaptive meshes are used as the
production is then not parallelised by Peano.
This might limit the scalability of the code.

Encountered a bug in dynamic AMR scenario.
This is not stable yet.
Will give it another revision while
I formalise it in my thesis.