Fix the re-coarsening bug where the mesh wasn't coarsened anymore. Thanks to JM for finding this one!
Revised the the a-posteriori refinement of the code:

We now keep the current solution, perform the mesh refinement,
and then perform a rollback afterwards.
This way we can refine the new solution multiple levels before doing
the rollback.

Furthermore, we can use the current and previous solution
in order to determine which cells to re-coarsen.

The re-coarsening was revised completely. It is now done via
a multi-step procedure which prevents refine-re-coarsening oscillations
and at the same time does not throw away high-resolution information
where those oscillations would occur:

First, if all children want to be erased, their solution and previous
solution is projected onto the coarse grid. Secondly, The refinement criterion
is evaluated for both solutions on the coarse grid. If any criteria requests refinement,
the re-coarsening procedure is stopped.
Thirdly, this decision is memorised and and the procedure will not start
again for this cell.

(Not sure if the MPI re-coarsening works yet if the tree is cut
at an inconvenient position. Do not have a test case.)