Optimised Riemann loops; obtained 1.1x SC speedup
Tested result with ApplicationExamples/TestCases/AMR_bug (should be renamed ;)).
Tested result with ApplicationExamples/Euler/Euler application
using 3 levels of AMR, limiter and fused time stepping.

I try to do the same optimisations to the MPI
part but it is currently pretty hard to get a
sufficiently big number of nodes to
run DDT on.