Added auxiliary Application "GRMHDb". This is a good interface between the "Ferrari-90" code, at the moment only for the GRMHD equations.
Michel accretion has been tested successfully in serial,
MPI does not work with the version of the 5 December 2018: further tests are needed.

The main file is GRMHD_mod.f90, it contains all the PDE.

I didn't commit the libtecio.a file, we should put it somewhere in the Miscellaneous folder.
(If needed, take it e.g. from .\ApplicationExamples\DIM\DIM_3DLADGOM\lib folder)