Carpet plotters: Preparing for immanent seperation from HDF5.
Big renaming and started to invent new plotter classes as a shim interface, "the exahype way", for the new Carpet+ASCII plotting alternatives. Anybody heard of orthogonality? Well, properly redoing the plotting API is something we have on the long-run list but I predict that during the lifetime of the project, nobody will rewrite all the plotter class hierarchy.