Toolkit: Adding flux-limiter choice for FV solvers.
Scope: This commit only affects the Abstract*Solvers and the kernels, the user API does not change.

Using this commit, users can choose now between different slope limiters (flux limiters) straight from the specification file.

Syntax is like:

  solver Finite-Volumes MyFVSolver
    variables const                = 172
    patch-size const               = 7
    maximum-mesh-size              = 1.2
    time-stepping                  = global
    type const         = musclhancock
    slope-limiter const  = koren
    terms const        = flux,ncp
    optimisation const = generic
    language const                 = C
    ...

Note the "slope-limiter" line.

Currently, the slope limiter is only used for the 2nd order MUSCL Hancock scheme. It is ignored for Godunov.

This commit is based on bc9063c5 and solves #274.